extract history building to new component
