+++ src/modules/archivist-api/components/ApiHistory/ApiHistory.tsx
@@ -1,38 +1,15 @@
-import { lighten, Typography, useTheme } from '@mui/material'
+import { Typography, useTheme } from '@mui/material'
 import { ButtonEx, FlexBoxProps, FlexCol } from '@xylabs/sdk-react'
 import { useState } from 'react'
 
-import { useArchivistApi } from '../../contexts'
 import { ApiCallEntry } from './ApiCallEntry'
 import { ApiCallTable } from './ApiCallTable'
+import { useBuildHistoryData } from './useBuildHistoryData'
 
 export const ApiHistory: React.FC<FlexBoxProps> = (props) => {
-  const { errorHistory, responseHistory, successHistory, failureHistory } = useArchivistApi()
   const [visible, setVisible] = useState(false)
   const theme = useTheme()
-
-  const histories = [
-    {
-      bgColor: lighten(theme.palette.success.light, 0.85),
-      callHistory: successHistory,
-      heading: 'Successes',
-    },
-    {
-      bgColor: lighten(theme.palette.error.light, 0.85),
-      callHistory: failureHistory,
-      heading: 'Failures',
-    },
-    {
-      bgColor: lighten(theme.palette.error.light, 0.85),
-      callHistory: errorHistory,
-      heading: 'Errors',
-    },
-    {
-      bgColor: lighten(theme.palette.info.light, 0.85),
-      callHistory: responseHistory,
-      heading: 'Responses',
-    },
-  ]
+  const histories = useBuildHistoryData()
 
   return (
     <FlexCol flexGrow={1} {...props}>
@@ -42,16 +19,16 @@ export const ApiHistory: React.FC<FlexBoxProps> = (props) => {
             Hide Api History
           </ButtonEx>
           {histories?.map((history, index) => (
-            <>
-              <Typography marginY={1} variant="h6" key={index}>
+            <FlexCol flexGrow={1} key={index}>
+              <Typography marginTop={1} variant="h6">
                 {history.heading}
               </Typography>
-              <ApiCallTable sx={{ marginBottom: theme.spacing(2) }}>
+              <ApiCallTable sx={{ marginBottom: theme.spacing(4) }}>
                 {history.callHistory?.map((response, index) => (
                   <ApiCallEntry bgColor={history.bgColor} call={response} index={index} key={index} />
                 ))}
               </ApiCallTable>
-            </>
+            </FlexCol>
           ))}
         </>
       ) : (
